requirements.txt
asyncio
discord-py-interactions
discord.py
requests